backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.17
google-generativeai==0.8.3
openpyxl==3.1.5
pandas==2.2.3
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.27.2
Pillow==11.1.0
